xaibio-backend/requirements.txt
annoy==1.17.3
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
facenet-pytorch==2.6.0
faiss-cpu==1.8.0
filelock==3.13.4
fonttools==4.53.0
fsspec==2024.3.1
gunicorn==20.1.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless==4.10.0.82
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-image==0.23.2
scipy==1.13.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tifffile==2024.5.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
